Implementing instructor CRUD

+++ src/main/java/com/comerlato/school_project/repository/DepartmentRepository.java
@@ -3,6 +3,6 @@ package com.comerlato.school_project.repository;
 import com.comerlato.school_project.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Long> {
+public interface DepartmentRepository extends JpaRepository<Department, Integer> {
 
 }
